requirements.txt
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
pandas==2.1.4
umap-learn==0.5.6